#48: moving JWT utility functions to separate file
