+++ utils.go
@@ -1,0 +1,53 @@
+package jwtware
+
+import (
+	"errors"
+	"strings"
+
+	"github.com/gofiber/fiber/v2"
+)
+
+// jwtFromHeader returns a function that extracts token from the request header.
+func jwtFromHeader(header string, authScheme string) func(c *fiber.Ctx) (string, error) {
+	return func(c *fiber.Ctx) (string, error) {
+		auth := c.Get(header)
+		l := len(authScheme)
+		if len(auth) > l+1 && strings.EqualFold(auth[:l], authScheme) {
+			return auth[l+1:], nil
+		}
+		return "", errors.New("Missing or malformed JWT")
+	}
+}
+
+// jwtFromQuery returns a function that extracts token from the query string.
+func jwtFromQuery(param string) func(c *fiber.Ctx) (string, error) {
+	return func(c *fiber.Ctx) (string, error) {
+		token := c.Query(param)
+		if token == "" {
+			return "", errors.New("Missing or malformed JWT")
+		}
+		return token, nil
+	}
+}
+
+// jwtFromParam returns a function that extracts token from the url param string.
+func jwtFromParam(param string) func(c *fiber.Ctx) (string, error) {
+	return func(c *fiber.Ctx) (string, error) {
+		token := c.Params(param)
+		if token == "" {
+			return "", errors.New("Missing or malformed JWT")
+		}
+		return token, nil
+	}
+}
+
+// jwtFromCookie returns a function that extracts token from the named cookie.
+func jwtFromCookie(name string) func(c *fiber.Ctx) (string, error) {
+	return func(c *fiber.Ctx) (string, error) {
+		token := c.Cookies(name)
+		if token == "" {
+			return "", errors.New("Missing or malformed JWT")
+		}
+		return token, nil
+	}
+}
